chore(server): rename geocodingController to geocodingService
