+++ packages/server/app/Controllers/Ws/MtvRoomsWsController.ts
@@ -5,7 +5,7 @@ import {
 } from '@musicroom/types';
 import MtvRoom from 'App/Models/MtvRoom';
 import User from 'App/Models/User';
-import GeocodingService from 'App/Services/GeocodingController';
+import GeocodingService from 'App/Services/GeocodingService';
 import SocketLifecycle from 'App/Services/SocketLifecycle';
 import UserService from 'App/Services/UserService';
 import { randomUUID } from 'crypto';
